Add files via upload
Please i need help. i am trying to dynamically include the products from the product array to the product section in the index page, but the images, title price and description are all stacking on each other. I am also trying to get the product id to be in the description.

Please help.